pyproject.toml
[project]
name = "chatbot-mvp"
version = "0.1.0"
description = "Knowledge base driven chatbot MVP"
requires-python = ">=3.10"

dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "faiss-cpu",              # 向量库
    "pypdf",                  # PDF 解析
    "python-docx",            # Word 文档解析
    "markdown",               # Markdown 解析
    "python-dotenv",          # 读取 .env
    "fastapi",                # Web API
    "uvicorn",                # 启动服务
    "dashscope",
    "langsmith",
    "langgraph",
    "python-multipart",
    "chromadb",
    "langchain_chroma"

]
